Refactor leader card utility
Refactor `MinistryRankingCurrentPartiesLeaderScoreboardChartsPageModContentFactoryImpl` to use `LeaderCardUtil` for creating leader cards and loading politicians.

* Add `LeaderCardUtil` class with methods for creating leader cards, loading politicians, and government role members.
* Remove duplicated methods for creating leader cards and loading politicians from `MinistryRankingCurrentPartiesLeaderScoreboardChartsPageModContentFactoryImpl`.
* Update dependencies in `MinistryRankingCurrentPartiesLeaderScoreboardChartsPageModContentFactoryImpl` to use `LeaderCardUtil`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Hack23/cia?shareId=XXXX-XXXX-XXXX-XXXX).
