+++ citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/goverment/pagemode/MinistryRankingCurrentPartiesLeaderScoreboardChartsPageModContentFactoryImpl.java
@@ -23,7 +23,7 @@ import com.hack23.cia.service.external.esv.api.EsvApi;
 import com.hack23.cia.service.external.esv.api.GovernmentBodyAnnualOutcomeSummary;
 import com.hack23.cia.service.external.esv.api.GovernmentBodyAnnualSummary;
 import com.hack23.cia.web.impl.ui.application.action.ViewAction;
-import com.hack23.cia.web.impl.ui.application.util.PartyLeaderUtil;
+import com.hack23.cia.web.impl.ui.application.util.LeaderCardUtil;
 import com.hack23.cia.web.impl.ui.application.views.common.pagemode.CardInfoRowUtil;
 import com.hack23.cia.web.impl.ui.application.views.common.pagemode.PoliticianLeaderboardUtil;
 import com.hack23.cia.web.impl.ui.application.views.common.rows.RowUtil;
@@ -71,6 +71,8 @@ public final class MinistryRankingCurrentPartiesLeaderScoreboardChartsPageModCon
 	@Autowired
 	private PoliticianLeaderboardUtil politicianLeaderboardUtil;
 
+	@Autowired
+	private LeaderCardUtil leaderCardUtil;
 
 	/**
 	 * Instantiates a new ministry ranking current parties leader scoreboard charts page mod content factory impl.
@@ -109,10 +111,10 @@ public final class MinistryRankingCurrentPartiesLeaderScoreboardChartsPageModCon
 		final ResponsiveRow row = RowUtil.createGridLayout(panelContent);
 		row.setSizeFull();
 
-		final List<ViewRiksdagenGovermentRoleMember> activeGovMembers = loadActiveGovernmentRoleMembers();
-		final Map<String, List<ViewRiksdagenPolitician>> activePoliticianMap = loadActivePoliticiansByPersonId();
+		final List<ViewRiksdagenGovermentRoleMember> activeGovMembers = leaderCardUtil.loadActiveGovernmentRoleMembers();
+		final Map<String, List<ViewRiksdagenPolitician>> activePoliticianMap = leaderCardUtil.loadActivePoliticiansByPersonId();
 
-		final Map<String, Boolean> partyLeaderMap = PartyLeaderUtil.computePartyLeaders(getApplicationManager(), activePoliticianMap.keySet());
+		final Map<String, Boolean> partyLeaderMap = leaderCardUtil.computePartyLeaders(activePoliticianMap.keySet());
 
 		// Sort roles
 		activeGovMembers.sort((a, b) -> {
@@ -138,7 +140,7 @@ public final class MinistryRankingCurrentPartiesLeaderScoreboardChartsPageModCon
 		    final ViewRiksdagenPoliticianExperienceSummary experienceSummary = getApplicationManager().getDataContainer(ViewRiksdagenPoliticianExperienceSummary.class).load(govMember.getPersonId());
 
 
-			final Panel cardPanel = createBaseballStyleCard(govMember, politician, ballotSummary, governmentBodyByMinistry,
+			final Panel cardPanel = leaderCardUtil.createBaseballStyleCard(govMember, politician, ballotSummary, governmentBodyByMinistry,
 					reportByMinistry,experienceSummary);
 
 			// Responsive column rules
@@ -150,31 +152,6 @@ public final class MinistryRankingCurrentPartiesLeaderScoreboardChartsPageModCon
 				NAME, parameters, pageId);
 
 		return panelContent;
-	}
-
-	/**
-	 * Load active government role members.
-	 *
-	 * @return the list
-	 */
-	private List<ViewRiksdagenGovermentRoleMember> loadActiveGovernmentRoleMembers() {
-		final DataContainer<ViewRiksdagenGovermentRoleMember, String> govermentRoleMemberDataContainer = getApplicationManager()
-				.getDataContainer(ViewRiksdagenGovermentRoleMember.class);
-		return govermentRoleMemberDataContainer.findListByProperty(new Object[] { Boolean.TRUE },
-				ViewRiksdagenGovermentRoleMember_.active);
-	}
-
-	/**
-	 * Load active politicians by person id.
-	 *
-	 * @return the map
-	 */
-	private Map<String, List<ViewRiksdagenPolitician>> loadActivePoliticiansByPersonId() {
-		final DataContainer<ViewRiksdagenPolitician, String> politicianDataContainer = getApplicationManager()
-				.getDataContainer(ViewRiksdagenPolitician.class);
-		final List<ViewRiksdagenPolitician> activePoliticians = politicianDataContainer
-				.findListByProperty(new Object[] { Boolean.TRUE }, ViewRiksdagenPolitician_.activeGovernment);
-		return activePoliticians.stream().collect(Collectors.groupingBy(ViewRiksdagenPolitician::getPersonId));
 	}
 
 	/**
@@ -197,167 +174,6 @@ public final class MinistryRankingCurrentPartiesLeaderScoreboardChartsPageModCon
 		} else {
 			return 5;
 		}
-	}
-
-	/**
-	 * Creates the baseball style card.
-	 *
-	 * @param govMember the gov member
-	 * @param politician the politician
-	 * @param ballotSummary the ballot summary
-	 * @param governmentBodyByMinistry the government body by ministry
-	 * @param reportByMinistry the report by ministry
-	 * @param experienceSummary
-	 * @return the panel
-	 */
-	private Panel createBaseballStyleCard(ViewRiksdagenGovermentRoleMember govMember,
-			ViewRiksdagenPolitician politician, ViewRiksdagenPoliticianBallotSummary ballotSummary, Map<String, List<GovernmentBodyAnnualSummary>> governmentBodyByMinistry,
-			Map<String, List<GovernmentBodyAnnualOutcomeSummary>> reportByMinistry, ViewRiksdagenPoliticianExperienceSummary experienceSummary) {
-
-		final Panel cardPanel = new Panel();
-		cardPanel.addStyleName("leader-baseball-card");
-		cardPanel.setSizeFull();
-		Responsive.makeResponsive(cardPanel);
-
-		final VerticalLayout cardContent = new VerticalLayout();
-		cardContent.setMargin(true);
-		cardContent.setSpacing(true);
-		cardContent.setSizeFull();
-		cardPanel.setContent(cardContent);
-
-		 CardInfoRowUtil.createCardHeader(cardContent,govMember.getRoleCode() + " " + govMember.getFirstName() + " " + govMember.getLastName()
-			+ " (" + govMember.getParty() + ")");
-		cardContent.addComponent(getPageLinkFactory().createPoliticianPageLink(politician));
-
-		final Link pageLink = new Link("Party " + politician.getParty(),
-				new ExternalResource("#!" + UserViews.PARTY_VIEW_NAME + "/" + politician.getParty()));
-		pageLink.setId(ViewAction.VISIT_PARTY_VIEW.name() + "/" + politician.getParty());
-		pageLink.setIcon(VaadinIcons.GROUP);
-
-		cardContent.addComponent(pageLink);
-
-		final boolean isPartyLeader = PartyLeaderUtil.isPartyLeader(getApplicationManager(), politician.getPersonId());
-		if (isPartyLeader) {
-			final ViewRiksdagenPartyRoleMember leaderRole = PartyLeaderUtil.getPartyLeaderRole(getApplicationManager(), politician.getPersonId());
-			if (leaderRole != null) {
-				final Label subHeader = new Label(
-						"Partiledare (" + govMember.getParty() + ") since " + leaderRole.getFromDate());
-				subHeader.addStyleName("card-subtitle");
-				cardContent.addComponent(subHeader);
-			}
-		}
-
-		// After creating the divider following the header/subtitle
-		// We create a vertical layout to hold Tenure and Experience on separate rows
-		final VerticalLayout statsContainer = new VerticalLayout();
-		statsContainer.setSpacing(false); // Less space between rows
-		statsContainer.addStyleName("card-stats-container");
-		statsContainer.setWidth("100%");
-
-		// Tenure Row
-		final HorizontalLayout tenureLayout = new HorizontalLayout();
-		tenureLayout.setSpacing(true);
-		tenureLayout.addStyleName("card-tenure");
-		final Label tenureIcon = new Label(VaadinIcons.CLOCK.getHtml(), ContentMode.HTML);
-		tenureIcon.setDescription("Tenure in days");
-
-		final Label tenureLabel = new Label("Tenure:");
-		tenureLabel.addStyleName("card-tenure-text");
-
-		final Label tenureValue = new Label(govMember.getTotalDaysServed() + " days");
-		tenureValue.addStyleName("card-tenure-value");
-
-		tenureLayout.addComponents(tenureIcon, tenureLabel, tenureValue);
-		statsContainer.addComponent(tenureLayout);
-
-		// Experience Row
-		final HorizontalLayout experienceLayout = new HorizontalLayout();
-		experienceLayout.setSpacing(true);
-		experienceLayout.addStyleName("card-experience-section");
-		final Label expIcon = new Label(VaadinIcons.USER_CHECK.getHtml(), ContentMode.HTML);
-		expIcon.setDescription("Political Experience");
-
-		final Label expLabel = new Label("Experience:");
-		expLabel.addStyleName("card-experience-text");
-
-		final int govYears = (int) (politician.getTotalDaysServedGovernment() / 365);
-		final int partyYears = (int) (politician.getTotalDaysServedParty() / 365);
-		final int parliamentYears = (int) (politician.getTotalDaysServedParliament() / 365);
-		final Label expValue = new Label(
-				"Goverment: " + govYears + "y, Party: " + partyYears + "y, Parliament: " + parliamentYears + "y");
-		expValue.addStyleName("card-experience-value");
-
-		experienceLayout.addComponents(expIcon, expLabel, expValue);
-		statsContainer.addComponent(experienceLayout);
-
-		// Add the statsContainer to the cardContent
-		cardContent.addComponent(statsContainer);
-
-		// Create grid for the four sections
-		final HorizontalLayout sectionsGrid = new HorizontalLayout();
-		sectionsGrid.setSpacing(true);
-		sectionsGrid.setWidth("100%");
-
-		// Add the four main sections
-		final VerticalLayout politicalRoleLayout = CardInfoRowUtil.createSectionLayout("Political Role & Influence");
-		addPoliticalRoleMetrics(politicalRoleLayout, govMember, politician, ballotSummary, experienceSummary);
-		sectionsGrid.addComponent(politicalRoleLayout);
-
-		final VerticalLayout performanceLayout = CardInfoRowUtil.createSectionLayout("Parliamentary Performance");
-		politicianLeaderboardUtil.addParliamentaryPerformanceMetrics(performanceLayout, politician, ballotSummary);
-		sectionsGrid.addComponent(performanceLayout);
-
-	    cardContent.addComponent(sectionsGrid);
-
-		final HorizontalLayout sections2Grid = new HorizontalLayout();
-		sections2Grid.setSpacing(true);
-		sections2Grid.setWidth("100%");
-
-
-		final VerticalLayout legislativeLayout = CardInfoRowUtil.createSectionLayout("Legislative Activity");
-		politicianLeaderboardUtil.addLegislativeMetrics(legislativeLayout, politician);
-		sections2Grid.addComponent(legislativeLayout);
-
-		final VerticalLayout alignmentLayout = CardInfoRowUtil.createSectionLayout("Party Alignment");
-		politicianLeaderboardUtil.addPartyAlignmentMetrics(alignmentLayout, politician, ballotSummary);
-		sections2Grid.addComponent(alignmentLayout);
-		cardContent.addComponent(sections2Grid);
-
-		politicianLeaderboardUtil.addMinistryRoleSummary(cardContent, govMember, governmentBodyByMinistry, reportByMinistry);
-
-		return cardPanel;
-	}
-
-
-	/**
-	 * Adds the political role metrics.
-	 *
-	 * @param layout the layout
-	 * @param govMember the gov member
-	 * @param politician the politician
-	 * @param ballotSummary the ballot summary
-	 * @param experienceSummary
-	 */
-	private void addPoliticalRoleMetrics(VerticalLayout layout, ViewRiksdagenGovermentRoleMember govMember,
-			ViewRiksdagenPolitician politician, ViewRiksdagenPoliticianBallotSummary ballotSummary, ViewRiksdagenPoliticianExperienceSummary experienceSummary) {
-
-		layout.addComponent(CardInfoRowUtil.createInfoRow("Current Role:", govMember.getRoleCode(), VaadinIcons.INSTITUTION,
-				"Current position in parliament"));
-			layout.addComponent(CardInfoRowUtil.createInfoRow("Career Length:",
-				String.format(Locale.ENGLISH,"%,d days", govMember.getTotalDaysServed()),
-				VaadinIcons.TIMER, "Years in parliament"));
-			layout.addComponent(CardInfoRowUtil.createInfoRow("Total Propositions:",
-					String.format(Locale.ENGLISH,"%,d", govMember.getTotalPropositions()),
-					VaadinIcons.GROUP, "Total Propositions"));
-			layout.addComponent(CardInfoRowUtil.createInfoRow("Total Government Bills:",
-					String.format(Locale.ENGLISH,"%,d", govMember.getTotalGovernmentBills()),
-					VaadinIcons.GROUP, "Total Government Bills"));
-
-			politicianLeaderboardUtil.addTopRoles(layout, experienceSummary);
-			politicianLeaderboardUtil.addKnowledgeAreas(layout, experienceSummary);
-			politicianLeaderboardUtil.addExperienceMetrics(layout,experienceSummary);
-			politicianLeaderboardUtil.addPoliticalAnalysisComment(layout, experienceSummary);
-
 	}
 
 	/**
