+++ citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/party/pagemode/PartyRankingCurrentPartiesLeaderScoreboardPageModContentFactoryImpl.java
@@ -41,6 +41,7 @@ import com.hack23.cia.service.external.esv.api.EsvApi;
 import com.hack23.cia.service.external.esv.api.GovernmentBodyAnnualOutcomeSummary;
 import com.hack23.cia.service.external.esv.api.GovernmentBodyAnnualSummary;
 import com.hack23.cia.web.impl.ui.application.action.ViewAction;
+import com.hack23.cia.web.impl.ui.application.util.LeaderCardUtil;
 import com.hack23.cia.web.impl.ui.application.util.PartyLeaderUtil;
 import com.hack23.cia.web.impl.ui.application.views.common.pagemode.CardInfoRowUtil;
 import com.hack23.cia.web.impl.ui.application.views.common.pagemode.PoliticianLeaderboardUtil;
@@ -79,6 +80,9 @@ public final class PartyRankingCurrentPartiesLeaderScoreboardPageModContentFacto
 
 	@Autowired
 	private PoliticianLeaderboardUtil politicianLeaderboardUtil;
+
+	@Autowired
+	private LeaderCardUtil leaderCardUtil;
 
 	/**
 	 * Instantiates a new party ranking current parties leader scoreboard page mod content factory impl.
@@ -119,7 +123,7 @@ public final class PartyRankingCurrentPartiesLeaderScoreboardPageModContentFacto
 		final ResponsiveRow row = RowUtil.createGridLayout(wrapper);
 		row.setSizeFull();
 
-		final Map<String, ViewRiksdagenPolitician> politicianMap = loadPoliticiansByPersonId();
+		final Map<String, ViewRiksdagenPolitician> politicianMap = leaderCardUtil.loadPoliticiansByPersonId(getApplicationManager());
 		final Map<String, Boolean> partyLeaderMap = PartyLeaderUtil.computePartyLeaders(getApplicationManager(), politicianMap.keySet());
 
 		final List<ViewRiksdagenPolitician> partyLeaders = politicianMap.values().stream()
@@ -152,7 +156,7 @@ public final class PartyRankingCurrentPartiesLeaderScoreboardPageModContentFacto
 					.getDataContainer(ViewRiksdagenPoliticianBallotSummary.class).load(leader.getPersonId());
 		    final ViewRiksdagenPoliticianExperienceSummary experienceSummary = getApplicationManager().getDataContainer(ViewRiksdagenPoliticianExperienceSummary.class).load(leader.getPersonId());
 
-			final Panel cardPanel = createLeaderCard(leader, ballotSummary, governmentBodyByMinistry, reportByMinistry,experienceSummary);
+			final Panel cardPanel = leaderCardUtil.createLeaderCard(getApplicationManager(), leader, ballotSummary, governmentBodyByMinistry, reportByMinistry, experienceSummary);
 			row.addColumn().withDisplayRules(12, 6, 4, 4).withComponent(cardPanel);
 		}
 
@@ -160,237 +164,6 @@ public final class PartyRankingCurrentPartiesLeaderScoreboardPageModContentFacto
 				parameters, pageId);
 
 		return panelContent;
-	}
-
-	/**
-	 * Load politicians by person id.
-	 *
-	 * @return the map
-	 */
-	private Map<String, ViewRiksdagenPolitician> loadPoliticiansByPersonId() {
-		final DataContainer<ViewRiksdagenPolitician, String> politicianDataContainer = getApplicationManager()
-				.getDataContainer(ViewRiksdagenPolitician.class);
-
-		final List<ViewRiksdagenPolitician> activePoliticians = politicianDataContainer.findListByProperty(
-				new Object[] { Boolean.TRUE }, ViewRiksdagenPolitician_.active);
-
-		return activePoliticians.stream().collect(Collectors.toMap(ViewRiksdagenPolitician::getPersonId, p -> p));
-	}
-
-	/**
-	 * Creates the leader card.
-	 *
-	 * @param leader the leader
-	 * @param ballotSummary the ballot summary
-	 * @param governmentBodyByMinistry the government body by ministry
-	 * @param reportByMinistry the report by ministry
-	 * @param experienceSummary
-	 * @return the panel
-	 */
-	private Panel createLeaderCard(final ViewRiksdagenPolitician leader, final ViewRiksdagenPoliticianBallotSummary ballotSummary,
-			final Map<String, List<GovernmentBodyAnnualSummary>> governmentBodyByMinistry,
-			final Map<String, List<GovernmentBodyAnnualOutcomeSummary>> reportByMinistry, final ViewRiksdagenPoliticianExperienceSummary experienceSummary) {
-
-		final Panel cardPanel = new Panel();
-		cardPanel.addStyleName("leader-baseball-card");
-		cardPanel.setSizeFull();
-		Responsive.makeResponsive(cardPanel);
-
-		final VerticalLayout cardContent = new VerticalLayout();
-		cardContent.setMargin(true);
-		cardContent.setSpacing(true);
-		cardContent.setSizeFull();
-		cardPanel.setContent(cardContent);
-
-		CardInfoRowUtil.createCardHeader(cardContent,"Partiledare " + leader.getFirstName() + " " + leader.getLastName() + " ("
-				+ leader.getParty() + ")");
-
-		// Politician detail link
-		cardContent.addComponent(getPageLinkFactory().createPoliticianPageLink(leader));
-
-		// Party link
-		final Link partyLink = new Link("Party " + leader.getParty(),
-				new ExternalResource("#!" + UserViews.PARTY_VIEW_NAME + "/" + leader.getParty()));
-		partyLink.setIcon(VaadinIcons.GROUP);
-		cardContent.addComponent(partyLink);
-
-		final boolean isPartyLeader = PartyLeaderUtil.isPartyLeader(getApplicationManager(), leader.getPersonId());
-		if (isPartyLeader) {
-			final ViewRiksdagenPartyRoleMember leaderRole = PartyLeaderUtil.getPartyLeaderRole(getApplicationManager(), leader.getPersonId());
-			if (leaderRole != null) {
-				final Label subHeader = new Label("Partiledare (" + leader.getParty() + ") since " + leaderRole.getFromDate());
-				subHeader.addStyleName("card-subtitle");
-				cardContent.addComponent(subHeader);
-			}
-		}
-
-		// Government or not
-		ViewRiksdagenGovermentRoleMember govMember = null;
-		if (leader.isActiveGovernment()) {
-			final Label govLabel = new Label("Currently in Government");
-			govLabel.addStyleName("card-subtitle");
-			cardContent.addComponent(govLabel);
-
-			// Add ministry summary if we can identify their ministry
-			// The ministry detail is stored in the same structure as the ministry snippet:
-			// We need to find which ministry they belong to
-			// In the ministry snippet, "govMember.getDetail()" gives ministry detail key.
-			// Here we only have leader, not govMember. We must find a corresponding approach:
-
-			// Let's assume we can identify the leader's ministry from active government roles data:
-			// we do similar approach: load active government role members and find the one matching this leader
-			final ViewRiksdagenPolitician pol = leader; // same as leader
-			govMember = findGovernmentRoleForLeader(pol);
-			if (govMember != null) {
-				politicianLeaderboardUtil.addMinistryRoleSummary(cardContent, govMember, governmentBodyByMinistry, reportByMinistry);
-			}
-
-		} else {
-			final Label nonGovLabel = new Label("Not in Government");
-			nonGovLabel.addStyleName("card-subtitle-nongov");
-			cardContent.addComponent(nonGovLabel);
-		}
-
-		// Tenure and Experience rows
-		final VerticalLayout statsContainer = new VerticalLayout();
-		statsContainer.setSpacing(false);
-		statsContainer.addStyleName("card-stats-container");
-		statsContainer.setWidth("100%");
-
-		// Tenure (assuming leader might have totalDaysServed property)
-		final Label tenureIcon = new Label(VaadinIcons.CLOCK.getHtml(), ContentMode.HTML);
-		tenureIcon.setDescription("Total Tenure");
-		final Label tenureLabel = new Label("Tenure:");
-		tenureLabel.addStyleName("card-tenure-text");
-		final Label tenureValue = new Label(leader.getTotalDaysServedParty() + " days");
-		tenureValue.addStyleName("card-tenure-value");
-		final HorizontalLayout tenureLayout = new HorizontalLayout(tenureIcon, tenureLabel, tenureValue);
-		tenureLayout.setSpacing(true);
-		tenureLayout.addStyleName("card-tenure");
-		statsContainer.addComponent(tenureLayout);
-
-		// Experience
-		final HorizontalLayout experienceLayout = new HorizontalLayout();
-		experienceLayout.setSpacing(true);
-		experienceLayout.addStyleName("card-experience-section");
-		final Label expIcon = new Label(VaadinIcons.USER_CHECK.getHtml(), ContentMode.HTML);
-		expIcon.setDescription("Political Experience");
-		final Label expLabel = new Label("Experience:");
-		expLabel.addStyleName("card-experience-text");
-
-		final int govYears = (int) (leader.getTotalDaysServedGovernment() / 365);
-		final int partyYears = (int) (leader.getTotalDaysServedParty() / 365);
-		final int parliamentYears = (int) (leader.getTotalDaysServedParliament() / 365);
-		final String expText = String.format(Locale.ENGLISH, "Government: %dy, Party: %dy, Parliament: %dy",
-				govYears, partyYears, parliamentYears);
-		final Label expValue = new Label(expText);
-		expValue.addStyleName("card-experience-value");
-		experienceLayout.addComponents(expIcon, expLabel, expValue);
-		statsContainer.addComponent(experienceLayout);
-
-		cardContent.addComponent(statsContainer);
-
-		// Create grid for the four sections
-		final HorizontalLayout sectionsGrid = new HorizontalLayout();
-		sectionsGrid.setSpacing(true);
-		sectionsGrid.setWidth("100%");
-
-
-		// Add the four main sections
-		final VerticalLayout politicalRoleLayout = CardInfoRowUtil.createSectionLayout("Political Role & Influence");
-		addPoliticalRoleMetrics(politicalRoleLayout, PartyLeaderUtil.getPartyLeaderRole(getApplicationManager(), leader.getPersonId()), govMember, leader, ballotSummary,experienceSummary);
-		sectionsGrid.addComponent(politicalRoleLayout);
-
-		final VerticalLayout performanceLayout = CardInfoRowUtil.createSectionLayout("Parliamentary Performance");
-		politicianLeaderboardUtil.addParliamentaryPerformanceMetrics(performanceLayout, leader, ballotSummary);
-		sectionsGrid.addComponent(performanceLayout);
-
-	    cardContent.addComponent(sectionsGrid);
-
-		final HorizontalLayout sections2Grid = new HorizontalLayout();
-		sections2Grid.setSpacing(true);
-		sections2Grid.setWidth("100%");
-
-
-		final VerticalLayout legislativeLayout = CardInfoRowUtil.createSectionLayout("Legislative Activity");
-		politicianLeaderboardUtil.addLegislativeMetrics(legislativeLayout, leader);
-		sections2Grid.addComponent(legislativeLayout);
-
-		final VerticalLayout alignmentLayout = CardInfoRowUtil.createSectionLayout("Party Alignment");
-		politicianLeaderboardUtil.addPartyAlignmentMetrics(alignmentLayout, leader, ballotSummary);
-		sections2Grid.addComponent(alignmentLayout);
-		cardContent.addComponent(sections2Grid);
-
-		return cardPanel;
-	}
-
-	/**
-	 * Find government role for leader.
-	 *
-	 * @param leader the leader
-	 * @return the view riksdagen goverment role member
-	 */
-	private ViewRiksdagenGovermentRoleMember findGovernmentRoleForLeader(ViewRiksdagenPolitician leader) {
-		// Similar to ministry snippet loadActiveGovernmentRoleMembers:
-		final DataContainer<ViewRiksdagenGovermentRoleMember, String> govermentRoleMemberDataContainer = getApplicationManager()
-				.getDataContainer(ViewRiksdagenGovermentRoleMember.class);
-		final List<ViewRiksdagenGovermentRoleMember> activeGovMembers = govermentRoleMemberDataContainer.findListByProperty(
-				new Object[] { Boolean.TRUE }, ViewRiksdagenGovermentRoleMember_.active);
-
-		// Find the government role that matches this leader's personId
-		return activeGovMembers.stream()
-				.filter(govMember -> govMember.getPersonId().equals(leader.getPersonId()))
-				.findFirst().orElse(null);
-	}
-
-
-
-	/**
-	 * Adds the political role metrics.
-	 *
-	 * @param layout the layout
-	 * @param riksdagenPartyRoleMember the riksdagen party role member
-	 * @param govMember the gov member
-	 * @param politician the politician
-	 * @param ballotSummary the ballot summary
-	 * @param experienceSummary
-	 */
-	private void addPoliticalRoleMetrics(VerticalLayout layout, ViewRiksdagenPartyRoleMember riksdagenPartyRoleMember, ViewRiksdagenGovermentRoleMember govMember,
-			ViewRiksdagenPolitician politician, ViewRiksdagenPoliticianBallotSummary ballotSummary, ViewRiksdagenPoliticianExperienceSummary experienceSummary) {
-
-		addPartyExperince(layout, riksdagenPartyRoleMember, govMember, politician);
-
-        // Top Roles
-		politicianLeaderboardUtil.addTopRoles(layout, experienceSummary);
-
-		// Top Knowledge Areas
-		politicianLeaderboardUtil.addKnowledgeAreas(layout, experienceSummary);
-
-		politicianLeaderboardUtil.addExperienceMetrics(layout,experienceSummary);
-
-		politicianLeaderboardUtil.addPoliticalAnalysisComment(layout, experienceSummary);
-
-	}
-
-	private void addPartyExperince(VerticalLayout layout, ViewRiksdagenPartyRoleMember riksdagenPartyRoleMember,
-			ViewRiksdagenGovermentRoleMember govMember, ViewRiksdagenPolitician politician) {
-		if (govMember != null) {
-			layout.addComponent(CardInfoRowUtil.createInfoRow("Role:", govMember != null ? govMember.getRoleCode() : "N/A", VaadinIcons.INSTITUTION,
-					"Current position in Government"));
-			layout.addComponent(CardInfoRowUtil.createInfoRow("Career Length Government:",
-					String.format(Locale.ENGLISH,"%,d days", govMember != null ? govMember.getTotalDaysServed() : 0),
-					VaadinIcons.TIMER, "Years in Government"));
-	    } else {
-			layout.addComponent(CardInfoRowUtil.createInfoRow("Career Length Parlimanet:",
-					String.format(Locale.ENGLISH,"%,d days", politician != null ? politician.getTotalDaysServedParliament() : 0),
-					VaadinIcons.TIMER, "Years in Parlimanet"));
-	    }
-
-		layout.addComponent(CardInfoRowUtil.createInfoRow("Current Party Role:", riksdagenPartyRoleMember != null ? riksdagenPartyRoleMember.getRoleCode() : "N/A", VaadinIcons.INSTITUTION,
-				"Current position in Party"));
-		layout.addComponent(CardInfoRowUtil.createInfoRow("Career Length Party Leader:",
-				String.format(Locale.ENGLISH,"%,d days", riksdagenPartyRoleMember != null ? riksdagenPartyRoleMember.getTotalDaysServed() : 0),
-				VaadinIcons.TIMER, "Years as Party Leader"));
 	}
 
 	/**
